Move fixture to test dir
